requirements.txt
attrs==25.3.0
certifi==2025.4.26
h11==0.16.0
idna==3.10
load-dotenv==0.1.0
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.1.0
selenium==4.32.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
